These changes provide advanced usage of bcachefs formatting options.
Because bcachefs provides
a one-stop-shop for wiping, partitioning, formatting and mounting with unique features for
each step, it has been broken out into it's own type if users wish to go beyond a single
disk setup.

examples/bcachefs.nix has been updated to showcase this advanced usage and some of the
available options.

Because all three processes take place via the type object, requiring "disks" attribute set
is not longer a hard requirement if a bcachefs type is provided. A check has been added to
account for that use case.